feat: allow to pass a slice to the preload function
I believe it's more idiomatic for a function to take a slice as a
parameter (vs. a pointer to a slice) to indicate that this function only
mutates the values in the underlying array, and not the length or the
capacity.

For example, that's the case for the sort functions in the standard
library.