feat: don't start a goroutine to call log hook
I think it's better to let the log hook decide if it wants to log
asynchronously or not.

The main motivation for this PR is that we have tests where we count the
number of sql queries via the log hook, and these tests are flaky
because the log hook is asynchronous.