ref: query directly, instead of preparing and closing statements
Currently, we always prepare, execute and close sql statements (without
reusing statements), which makes 3 roundtrips to the sql server.

Also, note that the sql driver at github.com/lib/pq does not implement
database/sql/drver.ConnPrepareContext, which means we still make a
roundtrip to the server for preparing the statement, even if the context
is cancelled.

With this PR, we query the server directly with only one roundtrip,
which is a bit more efficient.

This change makes instrumenting the underlying `*sql.DB` at the
driver level a bit simpler too, as we are not forced to handle the
database/sql/driver.ConnPrepareContext case.